pyproject.toml
[project]
name = "hugflow"
version = "0.1.0"
description = "GitOps-based declarative automation system for managing Hugging Face datasets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "huggingface-hub>=0.23.0",
    "PyGithub>=2.0.0",
    "structlog>=24.0.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.8.0",
]

[project.scripts]
hugflow = "hugflow.cli:app"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=hugflow --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
